chore(stores): move alt newRequestKey implementations to bench file
