+++ stores.go
@@ -3,9 +3,7 @@ package sdbc
 import (
 	"bytes"
 	cryptorand "crypto/rand"
-	"encoding/base64"
 	"encoding/binary"
-	"fmt"
 	"math/rand/v2"
 	"sync"
 )
@@ -74,7 +72,7 @@ type requests struct {
 }
 
 func (r *requests) prepare() (string, <-chan *output) {
-	key := newRequestKey4()
+	key := newRequestKey()
 	outChan := make(chan *output)
 
 	r.mut.Lock()
@@ -247,26 +245,9 @@ func (rb *RandBytes) Base62Str(length int) string {
 	return str
 }
 
-// Uniform distribution, but slower and variable key length (<= RequestKeyLength).
-func newRequestKey5() string {
-	key := make([]byte, RequestKeyLength)
-	randBytes.Read(key)
-
-	offs := 0
-	for i, b := range key {
-		if b > unbiasedMaxVal {
-			offs--
-			continue
-		}
-		key[i+offs] = charset[int(b)%charsetLen]
-	}
-
-	return string(key[:len(key)+offs])
-}
-
 // Fastest, but random distribution is not uniform.
 // Not security-critical in this case, so acceptable.
-func newRequestKey4() string {
+func newRequestKey() string {
 	key := make([]byte, RequestKeyLength)
 	randBytes.Read(key)
 
@@ -275,28 +256,4 @@ func newRequestKey4() string {
 	}
 
 	return string(key)
-}
-
-// Similar to official driver.
-func newRequestKey2() string {
-	return randBytes.Base62Str(RequestKeyLength)
-}
-
-// Using simpler rng, and base64.
-func newRequestKey1() string {
-	key := make([]byte, RequestKeyLength)
-	randBytes.Read(key)
-
-	return base64.RawURLEncoding.EncodeToString(key)
-}
-
-// Original uuid-like implementation.
-func newRequestKey0() string {
-	key := make([]byte, RequestKeyLength)
-
-	if _, err := cryptorand.Read(key); err != nil {
-		return "" // TODO: error?
-	}
-
-	return fmt.Sprintf("%X-%X-%X-%X-%X", key[0:4], key[4:6], key[6:8], key[8:10], key[10:])
 }
